requirements.txt
flask
torch
transformers
pillow
requests
opencv-python
numpy
